requirements.txt
praw>=7.8.1
prawcore>=2.4,<3

python-dotenv>=1.1.1
requests>=2.32.4
fastapi>=0.116.1

pydantic>=2.11.7
pydantic-settings>=2.10.1
pydantic-extra-types>=2.10.5
uvicorn>=0.35.0

Flask>=3.1.1
Flask-WTF>=1.2.2
Flask-SQLAlchemy>=3.1.1
flask-cors>=6.0.1
Flask-Toastr>=0.5.8
WTForms>=3.2.1
markdownify>=1.1.0

rq-dashboard-fast>=0.5.12
redis>=6.2.0
rq>=2.4.0
rq-scheduler>=0.14.0
alembic>=1.16.4
